+++ Parallel.Core/Storage/SshStorageProvider.cs
@@ -82,13 +82,16 @@ namespace Parallel.Core.Storage
         }
 
         /// <inheritdoc />
-        public async Task DownloadFileAsync(LocalFile file, string remotePath, CancellationToken ct = default)
+        public async Task<RemoteFile?> DownloadFileAsync(LocalFile file, string remotePath, CancellationToken ct = default)
         {
             InsureConnection();
+            if (!await ExistsAsync(remotePath)) return null;
+            
             await using SftpFileStream openStream = _client.OpenRead(remotePath);
             await using FileStream createStream = File.Create(file.Fullname);
             await using ZstdStream zstdStream = new(openStream, ZstdStreamMode.Decompress);
             await zstdStream.CopyToAsync(createStream, ct);
+            return await GetFileAsync(remotePath);
         }
 
         /// <inheritdoc />
